- 调整代码

+++ Engine/Resource/src/ResourceLoaderTexture.cpp
@@ -4,13 +4,13 @@
 #include "FileManager.h"
 
 
-static bool InternalLoadTexture2D(CGfxTexture2D *pTexture2D, const gli::texture *texture, int baseLevel)
+static bool InternalLoadTexture2D(CGfxTexture2D *pTexture2D, const gli::texture *texture, uint32_t baseLevel)
 {
 	gli::gl GL(gli::gl::PROFILE_ES30);
 	gli::gl::format format = GL.translate(texture->format(), texture->swizzles());
 
-	if (baseLevel > (int)texture->levels() - 1) {
-		baseLevel = (int)texture->levels() - 1;
+	if (baseLevel > texture->levels() - 1) {
+		baseLevel = texture->levels() - 1;
 	}
 
 	if (pTexture2D->Create((GfxPixelFormat)texture->format(), texture->extent(baseLevel).x, texture->extent(baseLevel).y, (int)texture->levels() - baseLevel) == false) {
@@ -35,13 +35,13 @@ static bool InternalLoadTexture2D(CGfxTexture2D *pTexture2D, const gli::texture 
 	return true;
 }
 
-static bool InternalLoadTexture2DArray(CGfxTexture2DArray *pTexture2DArray, const gli::texture *texture, int baseLevel)
+static bool InternalLoadTexture2DArray(CGfxTexture2DArray *pTexture2DArray, const gli::texture *texture, uint32_t baseLevel)
 {
 	gli::gl GL(gli::gl::PROFILE_ES30);
 	gli::gl::format format = GL.translate(texture->format(), texture->swizzles());
 
-	if (baseLevel > (int)texture->levels() - 1) {
-		baseLevel = (int)texture->levels() - 1;
+	if (baseLevel > texture->levels() - 1) {
+		baseLevel = texture->levels() - 1;
 	}
 
 	if (pTexture2DArray->Create((GfxPixelFormat)texture->format(), texture->extent(baseLevel).x, texture->extent(baseLevel).y, (int)texture->levels() - baseLevel, (int)texture->layers()) == false) {
@@ -70,13 +70,13 @@ static bool InternalLoadTexture2DArray(CGfxTexture2DArray *pTexture2DArray, cons
 	return true;
 }
 
-static bool InternalLoadTexture2DArrayLayer(CGfxTexture2DArray *pTexture2DArray, const gli::texture *texture, int layer, int baseLevel)
+static bool InternalLoadTexture2DArrayLayer(CGfxTexture2DArray *pTexture2DArray, const gli::texture *texture, int layer, uint32_t baseLevel)
 {
 	gli::gl GL(gli::gl::PROFILE_ES30);
 	gli::gl::format format = GL.translate(texture->format(), texture->swizzles());
 
-	if (baseLevel > (int)texture->levels() - 1) {
-		baseLevel = (int)texture->levels() - 1;
+	if (baseLevel > texture->levels() - 1) {
+		baseLevel = texture->levels() - 1;
 	}
 
 	if (gli::is_compressed(texture->format())) {
@@ -97,13 +97,13 @@ static bool InternalLoadTexture2DArrayLayer(CGfxTexture2DArray *pTexture2DArray,
 	return true;
 }
 
-static bool InternalLoadTextureCubeMap(CGfxTextureCubeMap *pTextureCubeMap, const gli::texture *texture, int baseLevel)
+static bool InternalLoadTextureCubeMap(CGfxTextureCubeMap *pTextureCubeMap, const gli::texture *texture, uint32_t baseLevel)
 {
 	gli::gl GL(gli::gl::PROFILE_ES30);
 	gli::gl::format format = GL.translate(texture->format(), texture->swizzles());
 
-	if (baseLevel > (int)texture->levels() - 1) {
-		baseLevel = (int)texture->levels() - 1;
+	if (baseLevel > texture->levels() - 1) {
+		baseLevel = texture->levels() - 1;
 	}
 
 	if (pTextureCubeMap->Create((GfxPixelFormat)texture->format(), texture->extent(baseLevel).x, texture->extent(baseLevel).y, (int)texture->levels() - baseLevel) == false) {
@@ -158,13 +158,13 @@ static bool InternalLoadTextureCubeMap(CGfxTextureCubeMap *pTextureCubeMap, cons
 	return true;
 }
 
-static bool InternalLoadTextureCubeMapFace(CGfxTextureCubeMap *pTextureCubeMap, const gli::texture *texture, GfxTextureCubeMapFace face, int baseLevel)
+static bool InternalLoadTextureCubeMapFace(CGfxTextureCubeMap *pTextureCubeMap, const gli::texture *texture, GfxTextureCubeMapFace face, uint32_t baseLevel)
 {
 	gli::gl GL(gli::gl::PROFILE_ES30);
 	gli::gl::format format = GL.translate(texture->format(), texture->swizzles());
 
-	if (baseLevel > (int)texture->levels() - 1) {
-		baseLevel = (int)texture->levels() - 1;
+	if (baseLevel > texture->levels() - 1) {
+		baseLevel = texture->levels() - 1;
 	}
 
 	if (gli::is_compressed(texture->format())) {
@@ -186,7 +186,7 @@ static bool InternalLoadTextureCubeMapFace(CGfxTextureCubeMap *pTextureCubeMap, 
 }
 
 
-bool CResourceLoader::LoadTexture2D(const char *szFileName, CGfxTexture2D *pTexture2D, int baseLevel)
+bool CResourceLoader::LoadTexture2D(const char *szFileName, CGfxTexture2D *pTexture2D, uint32_t baseLevel)
 {
 	pTexture2D->Destroy();
 	{
@@ -206,7 +206,7 @@ bool CResourceLoader::LoadTexture2D(const char *szFileName, CGfxTexture2D *pText
 	return false;
 }
 
-bool CResourceLoader::LoadTexture2DArray(const char *szFileName, CGfxTexture2DArray *pTexture2DArray, int baseLevel)
+bool CResourceLoader::LoadTexture2DArray(const char *szFileName, CGfxTexture2DArray *pTexture2DArray, uint32_t baseLevel)
 {
 	pTexture2DArray->Destroy();
 	{
@@ -226,7 +226,7 @@ bool CResourceLoader::LoadTexture2DArray(const char *szFileName, CGfxTexture2DAr
 	return false;
 }
 
-bool CResourceLoader::LoadTexture2DArrayLayer(const char *szFileName, int layer, CGfxTexture2DArray *pTexture2DArray, int baseLevel)
+bool CResourceLoader::LoadTexture2DArrayLayer(const char *szFileName, int layer, CGfxTexture2DArray *pTexture2DArray, uint32_t baseLevel)
 {
 	do {
 		CStream stream;
@@ -243,7 +243,7 @@ bool CResourceLoader::LoadTexture2DArrayLayer(const char *szFileName, int layer,
 	return false;
 }
 
-bool CResourceLoader::LoadTextureCubeMap(const char *szFileName, CGfxTextureCubeMap *pTextureCubeMap, int baseLevel)
+bool CResourceLoader::LoadTextureCubeMap(const char *szFileName, CGfxTextureCubeMap *pTextureCubeMap, uint32_t baseLevel)
 {
 	pTextureCubeMap->Destroy();
 	{
@@ -263,7 +263,7 @@ bool CResourceLoader::LoadTextureCubeMap(const char *szFileName, CGfxTextureCube
 	return false;
 }
 
-bool CResourceLoader::LoadTextureCubeMapFace(const char *szFileName, GfxTextureCubeMapFace face, CGfxTextureCubeMap *pTextureCubeMap, int baseLevel)
+bool CResourceLoader::LoadTextureCubeMapFace(const char *szFileName, GfxTextureCubeMapFace face, CGfxTextureCubeMap *pTextureCubeMap, uint32_t baseLevel)
 {
 	do {
 		CStream stream;
